Update app.py
Defined KNOWN_PARTICIPANTS at the top and I made a few changes on how the caller is considered.

When participants are from non-US numbers, or when making cross-country calls, the caller ID may not be preserved. That's why I trimmed the first few digits from the caller as seen by Twilio. Then check if the trimmed down caller ID is in the known participants dictionary. A false positive would be very unlikely and I think it's worth the risk.